pyproject.toml
[tool.poetry]
name = "symportal-kitchen"
version = "0.1.0"
description = "A collection of functions to maintain SymPortal Web Application and Computation Framework"
authors = ["greenjune-ship-it <yulia_iakovleva@proton.me>"]
license = "MIT"
readme = "README.md"
packages = [{include = "symportal_kitchen"}]

[tool.poetry.dependencies]
python = ">=3.8.1"
sendgrid = "^6.10.0"
python-dotenv = "^1.0.0"
jinja2 = "^3.1.2"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.3"
flake8 = "^6.0.0"
pylint = "^2.17.5"
mypy = "^1.4.1"
black = "^23.7.0"
autopep8 = "^2.0.2"
pytest = "^7.4.0"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
